Fix a SIGSEGV in olvwm
Use patch 0004-Avoid-stack-smashing-by-calling-GetWindowProperty-wi.patch
attached to https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=852532

From: =?UTF-8?q?Bernhard=20=C3=9Cbelacker?= <bernhardu@mailbox.org>
Date: Sat, 11 Mar 2017 15:21:39 +0100
Subject: Avoid stack smashing by calling GetWindowProperty with unsigned long
 instead unsigned int.

(gdb) bt
 #0  0x00007ffff76a0225 in XGetWindowProperty (dpy=dpy@entry=0x5555557a1040, w=669, property=349, offset=offset@entry=0, length=length@entry=33, delete=delete@entry=0, req_type=19, actual_type=0x7fffffffd9d0, actual_format=0x7fffffffd9c4, nitems=0x7fffffffda20, bytesafter=0x7fffffffda24, prop=0x7fffffffd9c8) at ../../src/GetProp.c:56
 #1  0x000055555556e5b6 in GetWindowProperty (dpy=dpy@entry=0x5555557a1040, w=<optimized out>, property=<optimized out>, long_offset=long_offset@entry=0, long_length=long_length@entry=33, req_type=req_type@entry=19, req_fmt=32, nitems=0x7fffffffda20, bytes_after=0x7fffffffda24) at properties.c:95
 #2  0x000055555555f2a3 in initClientState (dpy=0x5555557a1040) at client.c:186
 #3  InitClients (dpy=0x5555557a1040) at client.c:228
 #4  0x000055555555e7fd in main (argc=<optimized out>, argv=<optimized out>) at olwm.c:299

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=852532
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=855167
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=855168

+++ xview-3.2p1.4/clients/olwm/properties.c
@@ -507,7 +507,7 @@ PropGetWMState(dpy,win,state,iconwin)
 	int	*state;			/* RETURN */
 	Window	*iconwin;		/* RETURN */
 {
-	unsigned int nItems,remain;
+	unsigned long nItems,remain;
 	int	*data;
 
 	if (!PropAvailable(win,WMStateAvail))
@@ -568,7 +568,7 @@ PropGetOLWindowState(dpy,win,winState)
 	OLWindowState	*winState;		/* RETURN */
 {
 	OLWindowState	*newState;
-	unsigned int	nItems,remain;
+	unsigned long	nItems,remain;
 
 	if (!PropAvailable(win,OLWindowStateAvail))
 		return False;
